requirements.txt
uvicorn[standard]==0.29.0
httpx==0.27.0
mcp==1.9.2
slack_sdk==3.27.0
pydantic==2.7.2
google-api-python-client==2.171.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
tenacity==8.3.0
aiohttp==3.10.11
cryptography==45.0.3
asyncpg==0.30.0
